Adding a SparseMatrix class
Calculating a Neural Network involves a lot of matrix multiplication, and we get significant boosts for using a sparse matrix class. While there are definitely libraries out there, a large part of this project was a programming challenge for myself, so writing my own class seemed appropriate. Default storage is coordinate-wise storage, but will convert on-the-fly to CSC or CSR if that'll help optimize the algorithm.