Fixing bug with RDM openers
Intuition/experience told me that you Acceleration at T-10 seconds. That's largely true, but the status itself only lasts 10 seconds. This "works" in practice because you're actually hitting it 100-300ms later, in responce to the countdown timer, so that it's just-barely remaining when your spell goes off. Modifying the opener rotations to account for this.

So why wasn't this caught earlier? rand(), which was probably unseeded in the testcase toolkit, was triggering the 50/50 proc! So that was masking the issue all along.