Implementing the NEAT engine (#12)
* Uploading starting framework of NEST algorithm

Unlike the upload of the RDM Model, which was 100% feature-complete and unit-tested on upload, this is basically just a shell. Gotta start somewhere!

* Re-removing NEST in favor of NEAT

Seems to have lingered as part of a branch push error.

* Implement the NEAT connection gene

This class is the connection between nodes in our neural network.

* Implement the NEST node gene

This class is the nodes that comprise a neural network.

* Adding a SparseMatrix class

Calculating a Neural Network involves a lot of matrix multiplication, and we get significant boosts for using a sparse matrix class. While there are definitely libraries out there, a large part of this project was a programming challenge for myself, so writing my own class seemed appropriate. Default storage is coordinate-wise storage, but will convert on-the-fly to CSC or CSR if that'll help optimize the algorithm.

* Implement the NEAT Genome class

The genome is comprised of input/hidden/output nodes and weighted directioned connections. Built up of previously-submitted Connection and Node classes. Can generate sparse matrices of itself involved in matrix multiplication / fitness evaluation. Can generate the delta between itself an another genome (used in speciation).

* Add support for creating a new genome offspring

Crossover. Mutation. Adding random nodes and connections. Anything you could possibly want. It's exposed as a single public method on the Genome class, which'll make this a breeze to use.

* Removing get/setValue from Node class

We track node values in the sparse matrices. This field just isn't used.

* Implementing the NEAT Population class

Huzzah! This has a lot of power behind it, being responsible for tracking species lists, sorting new genomes into the appropriate species, and of course, taking the existing generation and generating a new one. Powerful!

* Fixed memleak in NEAT core classes

My raw pimpl pointers were causing a pretty fast memory leak. I was worried that I wouldn't be able to use std::unique_ptr since it was giving me a dll warning when I tried to dllexport the entire class. Solution is simply to dllexport every public method *instead* of the entire class, so that the unique_ptr doesn't get exported. Hooray for being able to use smart pointers!!!

* Improvement to RNG functions

My old methods used C-style rand() and probably also had a bug in randDouble(). Using the <random> header leads to much better output.

* Fixed my understanding of Wxh and Whh matrices

My original understanding of these connection matrices was confused and assumed that, like, Wxh was all forward connections and Whh was for anything considered recurrent. If my system was laid out differently, maybe this would make sense. But considering that my "input" is really input, output, and hidden states, I can use the same matrices for both.

ex. For Whh, the input-input section would all get multiplied by 0 (unless there were recurrent connections back to the input), giving us just the hidden-hidden connections in our resultant vector!

* Adding ReLU and softmax functions

These will be used by consuming code in evaluating a neural network.

* Add support for stale species culling

If a species hasn't made any improvements to its fitness in (default value) 15 generations, we start applying a % penalty to the adjusted fitness, encouraging them to die off.

* Updated pseudocode

Corrected the things I've since learned

* Deleting unneeded files

Oops! I guess I had intended the genome and the population to be in separate files? Possibly that'd be a good refactor for later, but for now, it's fine as is.