Added support for stale species support
This feature was originally broken because it was being updated on the old generation and not passed along to the next. This fixes that issue, so now stagnant species will get choked out. Yes this does have the possibility of choking out the best current species, but it also has the probability of spreading its genes and reblooming in another species later on.