Feature xor testcase (#15)
* Basic XOR test case

This is added to main toolkit instead of the test toolkit... but this is a nice proof of concept that can find a solution to XOR given two bits of input.

Things to do:
- we're not actually transferring the stale species info from generation to generation
- finding a solution takes a LOT of generations... much more than the NEAT paper
- probably need to play around with static settings... and those might be project-specific i.e. RDM will need different settings

* Added support for stale species support

This feature was originally broken because it was being updated on the old generation and not passed along to the next. This fixes that issue, so now stagnant species will get choked out. Yes this does have the possibility of choking out the best current species, but it also has the probability of spreading its genes and reblooming in another species later on.

* Fixing XOR example

I was using softmax on an array of three 0's and a value... Silly! Softmax doesn't make any sense here. Not only does this give me more accurate results, but it also (usually) resolves in fewer generations! Also, there's no need in this situation (I think?) for adjusting delta_t

* Forgot to submit main