Implement selection and execution of moves
Piecing it all together! This actually runs and does stuff! A terrific milestone.

This is enough now that the code will create the population of starting genomes, run them through the RDM model for (current setting) 3 ingame minutes, record their potency-per-second, and use (the square of) that as a fitness function for NEAT evolution!