Feature rdm testcase (#16)
* Add RDM class

Basic functionality present, the core loop

* Revert "Add RDM class"

This reverts commit 154fdf02ac21910c1e4c4c6ee218621c2e6611ba.

* Fixing branch merge issues

* Add support for converting state into an input vector

Vector is needed as input for each loop of the neural network.

* Convert resultant state vector into the selected move

We softmax the array of output nodes, and find the selected action that the neural network thinks we should (try) to execute next.

* Implement selection and execution of moves

Piecing it all together! This actually runs and does stuff! A terrific milestone.

This is enough now that the code will create the population of starting genomes, run them through the RDM model for (current setting) 3 ingame minutes, record their potency-per-second, and use (the square of) that as a fitness function for NEAT evolution!