Readd delta_t resizing as the genome grows
With the way that the NEAT paper (Stanley, Miikkulainen) defines their delta_t function, there ends up being a large delta comparing small genomes and tiny delta comparing large genomes. As a consequence, after a few hundred generations, the population collapses down into a single species, reducing the chance for a new species to evolve. It's important to shrink delta_t as necessary to maintain a minimum level of diversity.